pyproject.toml
[project]
name = "a2a-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-adk>=1.18.0",
    "jinja2>=3.1.6",
    "rich>=13.7.1",
    "pytest>=8.3.2",
    "typing-extensions>=4.12.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
